main
Added a minor tweak in which the number of battles will be reduced in the possible event of a Optimus-Predaking match up.  Added a default null to TransformerImpl.  Fixed the Optional handling of TransformerServiceImpl.  Added more error handling to TransformerController to behave like a web interface should.  Adjusted TransformerControllerTest to recognize this new error handling.  Finished README.md.